+++ R/tar_resources_gcp.R
@@ -29,8 +29,11 @@
 #'   format = "qs",
 #'   repository = "gcp",
 #'   resources = tar_resources(
-#'     gcp = tar_resources_gcp(bucket = "yourbucketname"),
-#'     qs = tar_resources_qs(preset = "fast")
+#'     gcp = tar_resources_gcp(
+#'       bucket = "yourbucketname",
+#'       prefix = "_targets"
+#'     ),
+#'     qs = tar_resources_qs(preset = "fast"),
 #'   )
 #' )
 tar_resources_gcp <- function(